- getting profile data from backend - sending logout request on logout - profile screen: pull to refresh + spent amount from backend - profile screen: opening terms and conditions/privacy url in browser - profile color: setting color from backend + color picker - profile photo: select from gallery - expenses on map: navigation on current location

+++ app/src/main/java/com/joerakhimov/smartexpenses/screen/main/profile/ProfileViewModel.kt
@@ -1,27 +1,71 @@
 package com.joerakhimov.smartexpenses.screen.main.profile
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
+import com.joerakhimov.smartexpenses.base.BaseViewModel
 import com.joerakhimov.smartexpenses.data.repository.SmartExpensesRepository
 import com.joerakhimov.smartexpenses.di.Injector
+import com.joerakhimov.smartexpenses.helper.concurrent.SchedulerProvider
+import com.joerakhimov.smartexpenses.screen.main.profile.model.ProfileInfo
 import javax.inject.Inject
 
-class ProfileViewModel: ViewModel() {
+class ProfileViewModel : BaseViewModel() {
 
     @Inject
     lateinit var repository: SmartExpensesRepository
 
+    @Inject
+    lateinit var schedulerProvider: SchedulerProvider
+
     val email = MutableLiveData<String>()
+
+    var profileInfo = MutableLiveData<ProfileInfo?>()
 
     init {
         Injector.appComponent.inject(this)
         email.value = repository.getEmail()
+        getProfileInfo()
+    }
+
+    private fun getProfileInfo() {
+        repository.getProfile()
+            .subscribeOn(schedulerProvider.io)
+            .observeOn(schedulerProvider.ui)
+            .doOnSubscribe { isLoading.set(true) }
+            .doFinally { isLoading.set(false) }
+            .subscribe({
+                if (it.status == 0) {
+                    if (it.messages?.size!! > 0 && it.messages[0] != null) {
+                        profileInfo.value = it.messages[0]
+                    }
+                } else toastMessage.value = it.message
+            }, {
+                toastMessage.value = it.message
+            })
     }
 
     fun onLogoutButtonClick() {
-        repository.removeToken()
-        repository.removeEmail()
+        repository.logout()
+            .subscribeOn(schedulerProvider.io)
+            .observeOn(schedulerProvider.ui)
+            .doOnSubscribe { isLoading.set(true) }
+            .doFinally { isLoading.set(false) }
+            .subscribe({
+                if (it.status == 0) {
+                    repository.removeToken()
+                    repository.removeEmail()
+                    isOpen.set(false)
+                } else toastMessage.value = it.message
+            }, {
+                toastMessage.value = it.message
+            })
+    }
+
+    fun onRefresh() {
+        getProfileInfo()
+    }
+
+    fun onColorSelection(selectedColor: String) {
+
     }
 
 }
